voice-ai-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0
websockets==13.0
google-genai==1.49.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.9.2
httpx==0.28.1
librosa==0.10.2
soundfile==0.12.1
